Much house keeping:
Follow the projects naming convention for the List<Instruction> that gets passed around everywhere.

Use IsIlLoop instead of a for loop.

Change IsLoop to IsIlLoop, move it to HelperFuncs.cs and make it internal instead of private.

Replace ProgramValidator.NullProgram(List<Instruction>) with a simple if check and remove an unneeded using statement.

Fix some documentation / comments

Reorganize Optimizer.cs and HelperFuncs.cs